+++ CHANGELOG.md
@@ -24,6 +24,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   for Console Metric Exporter.
 - Enable ASLR, CET and CFG for native profiler libraries on Windows.
 - Enable compiler flags to harden native profiler libraries on macOS and Linux.
+- Support for [MongoDB.Driver](https://www.nuget.org/packages/MongoDB.Driver/)
+  traces instrumentation for versions `3.5.0`+.
 
 ### Changed